Fix incorrect expectations on getReferenceKey without and right type specifier
These should be false, as `link.other.getReferenceKey()` should return a key for `Patient/p3`. The absence of the type specifier should not cause this to return an empty collection, rather the collection should be unconstrained by the type of the reference.

In the case where it is the right type specifier, it should also return a non-empty collection.